termine el diseño del home completo
