--- src/Components/Footer/Footer.module.css
@@ -1,57 +1,0 @@
-
-.Footer {
-  background-color: #D9D9D9;
-  color: #191919;
-  display: grid;
-  font-family: 'Barlow', sans-serif;
-  font-weight: bold;
-  width: 100vw;
-  height: 300px;
-  margin-left: -50px;
-  margin-bottom: -50px;
-}
-
-.mainContent {
-  flex: 1;  
-}
-
-.TeamSection {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  gap: 10px;
-} 
-
-.textoTeam {
-  margin-top: 72px;
-}
-
-
-ul {
-    list-style: none;
-} 
-.Footer a {
-  text-decoration: none;  
-  color: black;           
-}
-
-
-.Footer a:hover {
-  text-decoration: underline; 
-}
-
-.contactos{
-  text-align: right;
-  margin-right: 50px;
-  margin-top: 55px;
-}
-
-.FooterCopyright {
-  grid-column: span 3;
-  text-align: center;
-}
-
-.imagenFooter img {
-  height: 200px;
-  width: auto;
-  margin-top: 50px;
-}
