--- src/Components/Offer/Offer.module.css
@@ -1,32 +1,0 @@
-.offer {
-    background: #D9D9D9;
-    max-width: 1200px;
-    margin: 0 auto;
-}
-.offerBackground {
-  background: #D9D9D9; 
-  height: 450px;
-  /* margin-inline: -48px; */
-  margin: 0 auto;
-  position: relative;
-  left: 0;
-  right: 0;
-  top: -250px;
-}
-
-.cardOferta{
-  margin: 0 auto;
-  justify-content: center;
-}
-
-
-.offer h1{ 
-  color: #000;
-  font-family: Cinzel;
-  font-size: 40px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-  margin-top: 10px;
-}
-
