+++ src/views/Home/Home.module.css
@@ -1,98 +1,92 @@
 .home {
-    padding: 40px 40px 40px 40px;
-    justify-content: center;
-    align-items: center;
-    min-height: 100vh;
+  padding-inline: 40px;
+  justify-content: center;
 }
 
+/* Seccion 1 */
 
-.contentWrapper {
-    min-height: calc(100vh - [altura del footer]);
-  }
-
+.primerseccion {
+  padding-inline: 20px;
+  position: relative;
+  height: 500px;
+}
 .quienes h1 {
-    color: white;
-    font-family: Cinzel;
-    font-size: 60px;
-    font-style: normal;
-    font-weight: 400;
-    text-align: left;
-    margin-bottom: -50px;
-    margin-left: 40px;
+  color: white;
+  font-family: Cinzel;
+  font-size: 60px;
+  font-weight: 400;
+  text-align: left;
+  margin-bottom: -50px;
 }
 
-.home .quienes h1:first-child {
-    margin-bottom: -35px; 
-    font-size: 70px;    
+.quienes h1:first-child {
+  margin-bottom: -35px;
+  font-size: 70px;
 }
 
 .texto {
-    text-align: justify; 
-    display: block;      
-    margin-left: 40px;   
-    max-width: 600px;    
-    color: white;
-    margin-top: 70px;
-    font-family: 'Barlow', sans-serif;
+  display: block;
+  max-width: 600px;
+  text-align: left;
+  color: white;
+  margin-top: 70px;
+  font-family: "Barlow", sans-serif;
 }
 
-
 .repisas {
-    position: relative;
-    top: -300px;        /* Alinea al inicio de .home */
-    right: -350px;      /* Alinea a la derecha */
-    width: 300px;  /* Ajusta a tu necesidad */
-    height: 500px;  /* Mantiene la relación de aspecto */
+  position: absolute;
+  bottom: 0;
+  right: 0;
+  width: 300px; /* Ajusta a tu necesidad */
+  height: 500px; /* Mantiene la relación de aspecto */
 }
 
 .repisa {
-    position: relative;
-    top: -300px;        /* Alinea al inicio de .home */
-    right: -325px;      /* Alinea a la derecha */
-    width: 300px;  /* Ajusta a tu necesidad */
-    height: 500px;  /* Mantiene la relación de aspecto */
+  position: absolute;
+  bottom: 0;
+  right: 320px;
+  width: 300px; /* Ajusta a tu necesidad */
+  height: 500px; /* Mantiene la relación de aspecto */
 }
 
+/*  Section 3 */
+
 .contenedor {
-    display: flex;
-    flex-direction: column;
-    align-items: center; /* Centra los elementos horizontalmente */
-    justify-content: center; /* Centra los elementos verticalmente */
-    width: 100%; /* Ocupa todo el ancho disponible */
-    margin-top: 50px;
-    position: relative;
-    top: -200px;
+  position: inherit;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 600px;
 }
 
 .contenedor h1 {
-    text-align: center; /* Centra el texto de los títulos */
-    margin: 20px 0; /* Espacio arriba y abajo del título */
-    color: white;
-    font-family: Cinzel;
-    font-size: 40px;
-    font-style: normal;
-    font-weight: 400;
+  text-align: center;
+  color: white;
+  font-family: Cinzel;
+  font-size: 40px;
+  font-weight: 400;
+  margin: 20px auto;
 }
 
-.contenedor .servicio {
-    display: flex; /* Alinea la imagen y el título horizontalmente */
-    flex-direction: column; /* Organiza los elementos verticalmente */
-    align-items: center; /* Centra los elementos horizontalmente */
-    margin: 60px 90px; /* Espacio entre las imágenes/títulos */
+.servicios {
+  display: flex;
+  justify-content: center;
+  gap: 100px;
 }
 
-.contenedor .servicios {
-    display: flex; /* Alinea las imágenes/títulos horizontalmente */
-    justify-content: center; /* Centra los elementos */
-    width: 100%;
+.servicio {
+  display: flex; /* Alinea la imagen y el título horizontalmente */
+  flex-direction: column; /* Organiza los elementos verticalmente */
+  align-items: center;
 }
 
 .muebles {
-    height: auto;
-    width: 400px;
+  height: 400px;
+  width: 400px;
 }
 
-.decoracion {
-    height: auto;
-    width: 400px;
+.servicio h1 {
+  margin: 20px auto;
+  font-size: 30px;
 }
